refactor: move serialization helpers into one place
