+++ advanced_alchemy/config/engine.py
@@ -3,6 +3,7 @@ from __future__ import annotations
 from dataclasses import dataclass
 from typing import TYPE_CHECKING, Callable, Literal
 
+from advanced_alchemy._serialization import decode_json, encode_json
 from advanced_alchemy.config.types import Empty
 
 if TYPE_CHECKING:
@@ -14,26 +15,6 @@ if TYPE_CHECKING:
 
     from advanced_alchemy.config.types import EmptyType
 
-try:
-    from msgspec.json import Decoder, Encoder
-
-    encoder, decoder = Encoder(), Decoder()
-    decode_json = decoder.decode
-
-    def encode_json(data: Any) -> str:
-        return encoder.encode(data).decode("utf-8")
-
-except ImportError:
-    try:
-        from orjson import dumps as _encode_json
-        from orjson import loads as decode_json  # type: ignore[no-redef]
-
-        def encode_json(data: Any) -> str:
-            return _encode_json(data).decode("utf-8")  # type: ignore[no-any-return]
-
-    except ImportError:
-        from json import dumps as encode_json  # type: ignore[assignment]
-        from json import loads as decode_json  # type: ignore[assignment]
 
 __all__ = ("EngineConfig",)
